asset-list-screen is created. library-assets-screen is created.
