+++ app/screens/index.ts
@@ -1,5 +1,7 @@
 export * from "./home/home-screen"
 export * from "./library/library-screen"
+export * from "./library/library-assets-screen"
+export * from "./asset-list/asset-list-screen"
 export * from "./error/error-boundary"
 export * from "./photo"
 // export other screens here
