+++ app/store/index.ts
@@ -1,5 +1,5 @@
 import { atom } from "recoil"
-import { Asset,RecyclerAssetListSection } from "../types"
+import { Asset, RecyclerAssetListSection, Library } from "../types"
 
 export const mediasState = atom<Asset[]>({
   key: "mediasState",
@@ -8,5 +8,10 @@ export const mediasState = atom<Asset[]>({
 
 export const recyclerSectionsState = atom<RecyclerAssetListSection[]>({
   key: "recyclerSectionsState",
+  default: null,
+})
+
+export const selectedLibraryState = atom<Library>({
+  key: "selectedLibraryState",
   default: null,
 })
